device_driver

Struct CommandOperation

pub struct CommandOperation<'b, B, AddressType, InFieldset, OutFieldset, Repeat>
where
    B: Block,
    B::Interface: CommandInterfaceBase<AddressType = AddressType>,
    AddressType: Address,
{ /* private fields */ }

Intermediate type for doing command operations

Implementations

impl<'d, B, AddressType, InFieldset, OutFieldset, Repeat> CommandOperation<'d, B, AddressType, InFieldset, OutFieldset, Repeat>
where B: Block, B::Interface: CommandInterfaceBase<AddressType = AddressType>, AddressType: Address,

pub fn dispatch( self, ) -> Result<(), <B::Interface as CommandInterfaceBase>::Error>
where B::Interface: CommandInterface, InFieldset: NotFieldset, OutFieldset: NotFieldset, Repeat: NotRepeating,

Dispatch the command to the device

pub fn dispatch_at( self, index: Repeat::Index, ) -> Result<(), <B::Interface as CommandInterfaceBase>::Error>
where B::Interface: CommandInterface, InFieldset: NotFieldset, OutFieldset: NotFieldset, Repeat: Repeating,

Dispatch the command to the device

pub fn dispatch_async( self, ) -> impl Future<Output = Result<(), <B::Interface as CommandInterfaceBase>::Error>>
where B::Interface: AsyncCommandInterface, InFieldset: NotFieldset, OutFieldset: NotFieldset, Repeat: NotRepeating,

Dispatch the command to the device

pub fn dispatch_at_async( self, index: Repeat::Index, ) -> impl Future<Output = Result<(), <B::Interface as CommandInterfaceBase>::Error>>
where B::Interface: AsyncCommandInterface, InFieldset: NotFieldset, OutFieldset: NotFieldset, Repeat: Repeating,

Dispatch the command to the device

pub fn dispatch_in( self, f: impl FnOnce(&mut InFieldset), ) -> Result<(), <B::Interface as CommandInterfaceBase>::Error>
where B::Interface: CommandInterface, InFieldset: Fieldset, OutFieldset: NotFieldset, Repeat: NotRepeating,

Dispatch the command to the device with an input

pub fn dispatch_in_at( self, index: Repeat::Index, f: impl FnOnce(&mut InFieldset), ) -> Result<(), <B::Interface as CommandInterfaceBase>::Error>
where B::Interface: CommandInterface, InFieldset: Fieldset, OutFieldset: NotFieldset, Repeat: Repeating,

Dispatch the command to the device with an input

pub async fn dispatch_in_async( self, f: impl FnOnce(&mut InFieldset), ) -> Result<(), <B::Interface as CommandInterfaceBase>::Error>
where B::Interface: AsyncCommandInterface, InFieldset: Fieldset, OutFieldset: NotFieldset, Repeat: NotRepeating,

Dispatch the command to the device with an input

pub async fn dispatch_in_at_async( self, index: Repeat::Index, f: impl FnOnce(&mut InFieldset), ) -> Result<(), <B::Interface as CommandInterfaceBase>::Error>
where B::Interface: AsyncCommandInterface, InFieldset: Fieldset, OutFieldset: NotFieldset, Repeat: Repeating,

Dispatch the command to the device with an input

pub fn dispatch_out( self, ) -> Result<OutFieldset, <B::Interface as CommandInterfaceBase>::Error>
where B::Interface: CommandInterface, InFieldset: NotFieldset, OutFieldset: Fieldset, Repeat: NotRepeating,

Dispatch the command to the device with an output

pub fn dispatch_out_at( self, index: Repeat::Index, ) -> Result<OutFieldset, <B::Interface as CommandInterfaceBase>::Error>
where B::Interface: CommandInterface, InFieldset: NotFieldset, OutFieldset: Fieldset, Repeat: Repeating,

Dispatch the command to the device with an output

pub async fn dispatch_out_async( self, ) -> Result<OutFieldset, <B::Interface as CommandInterfaceBase>::Error>
where B::Interface: AsyncCommandInterface, InFieldset: NotFieldset, OutFieldset: Fieldset, Repeat: NotRepeating,

Dispatch the command to the device with an output

pub async fn dispatch_out_at_async( self, index: Repeat::Index, ) -> Result<OutFieldset, <B::Interface as CommandInterfaceBase>::Error>
where B::Interface: AsyncCommandInterface, InFieldset: NotFieldset, OutFieldset: Fieldset, Repeat: Repeating,

Dispatch the command to the device with an output

pub fn dispatch_inout( self, f: impl FnOnce(&mut InFieldset), ) -> Result<OutFieldset, <B::Interface as CommandInterfaceBase>::Error>
where B::Interface: CommandInterface, InFieldset: Fieldset, OutFieldset: Fieldset, Repeat: NotRepeating,

Dispatch the command to the device with an input and output

pub fn dispatch_inout_at( self, index: Repeat::Index, f: impl FnOnce(&mut InFieldset), ) -> Result<OutFieldset, <B::Interface as CommandInterfaceBase>::Error>
where B::Interface: CommandInterface, InFieldset: Fieldset, OutFieldset: Fieldset, Repeat: Repeating,

Dispatch the command to the device with an input and output

pub async fn dispatch_inout_async( self, f: impl FnOnce(&mut InFieldset), ) -> Result<OutFieldset, <B::Interface as CommandInterfaceBase>::Error>
where B::Interface: AsyncCommandInterface, InFieldset: Fieldset, OutFieldset: Fieldset, Repeat: NotRepeating,

Dispatch the command to the device with an input and output

pub async fn dispatch_inout_at_async( self, index: Repeat::Index, f: impl FnOnce(&mut InFieldset), ) -> Result<OutFieldset, <B::Interface as CommandInterfaceBase>::Error>
where B::Interface: AsyncCommandInterface, InFieldset: Fieldset, OutFieldset: Fieldset, Repeat: Repeating,

Dispatch the command to the device with an input and output