device_driver

Trait AsyncRegisterInterface

pub trait AsyncRegisterInterface: RegisterInterfaceBase {
    // Required methods
    async fn write_register(
        &mut self,
        address: Self::AddressType,
        data: &mut [u8],
        _metadata: &FieldsetMetadata,
    ) -> Result<(), Self::Error>;
    async fn read_register(
        &mut self,
        address: Self::AddressType,
        data: &mut [u8],
        _metadata: &FieldsetMetadata,
    ) -> Result<(), Self::Error>;
}

A trait to represent the interface to the device.

This is called to asynchronously write to and read from registers.

Required Methods

async fn write_register( &mut self, address: Self::AddressType, data: &mut [u8], _metadata: &FieldsetMetadata, ) -> Result<(), Self::Error>

Write the given data to the register located at the given address

async fn read_register( &mut self, address: Self::AddressType, data: &mut [u8], _metadata: &FieldsetMetadata, ) -> Result<(), Self::Error>

Read the register located at the given address to the given data slice

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".