device_driver

Trait BufferInterface

pub trait BufferInterface: BufferInterfaceBase {
    // Required methods
    fn write(
        &mut self,
        address: Self::AddressType,
        buf: &[u8],
    ) -> Result<usize, Self::Error>;
    fn flush(&mut self, address: Self::AddressType) -> Result<(), Self::Error>;
    fn read(
        &mut self,
        address: Self::AddressType,
        buf: &mut [u8],
    ) -> Result<usize, Self::Error>;
}

A trait to represent the interface to the device.

This is called to read from and write to buffers.

Required Methods

fn write( &mut self, address: Self::AddressType, buf: &[u8], ) -> Result<usize, Self::Error>

Write to the buffer with the given address.

This interface must adhere to [embedded_io::Write::write].

fn flush(&mut self, address: Self::AddressType) -> Result<(), Self::Error>

Flush this output stream with the given address.

This interface must adhere to [embedded_io::Write::flush].

fn read( &mut self, address: Self::AddressType, buf: &mut [u8], ) -> Result<usize, Self::Error>

Read from the buffer with the given address.

This interface must adhere to [embedded_io::Read::read].

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".