device_driver

Struct BufferOperation

pub struct BufferOperation<'i, Interface, AddressType: Copy, Access> { /* private fields */ }

Intermediate type for doing buffer operations

If the interface error implements embedded_io::Error, then this operation type also implements the embedded_io traits

Implementations

impl<Interface, AddressType: Copy, Access> BufferOperation<'_, Interface, AddressType, Access>
where Interface: BufferInterface<AddressType = AddressType>, Access: WriteCapability,

pub fn write(&mut self, buf: &[u8]) -> Result<usize, Interface::Error>

Write a buffer into this writer, returning how many bytes were written.

Mirror function of embedded_io::Write::write.

pub fn write_all(&mut self, buf: &[u8]) -> Result<(), Interface::Error>

Write an entire buffer into this writer.

This function calls write() in a loop until exactly buf.len() bytes have been written, blocking if needed.

Mirror function of embedded_io::Write::write_all.

pub fn flush(&mut self) -> Result<(), Interface::Error>

Flush this output stream, blocking until all intermediately buffered contents reach their destination.

Mirror function of embedded_io::Write::flush.

impl<Interface, AddressType: Copy, Access> BufferOperation<'_, Interface, AddressType, Access>
where Interface: BufferInterface<AddressType = AddressType>, Access: ReadCapability,

pub fn read(&mut self, buf: &mut [u8]) -> Result<usize, Interface::Error>

Read some bytes from this source into the specified buffer, returning how many bytes were read.

Mirror function of embedded_io::Read::read.

pub fn read_exact( &mut self, buf: &mut [u8], ) -> Result<(), ReadExactError<Interface::Error>>

Read the exact number of bytes required to fill buf. This function calls read() in a loop until exactly buf.len() bytes have been read, blocking if needed.

Mirror function of embedded_io::Read::read_exact.

impl<Interface, AddressType: Copy, Access> BufferOperation<'_, Interface, AddressType, Access>
where Interface: AsyncBufferInterface<AddressType = AddressType>, Access: WriteCapability,

pub async fn write_async( &mut self, buf: &[u8], ) -> Result<usize, Interface::Error>

Write a buffer into this writer, returning how many bytes were written.

Mirror function of embedded_io_async::Write::write.

pub async fn write_all_async( &mut self, buf: &[u8], ) -> Result<(), Interface::Error>

Write an entire buffer into this writer.

This function calls write() in a loop until exactly buf.len() bytes have been written, blocking if needed.

Mirror function of embedded_io_async::Write::write_all.

pub async fn flush_async(&mut self) -> Result<(), Interface::Error>

Flush this output stream, blocking until all intermediately buffered contents reach their destination.

Mirror function of embedded_io_async::Write::flush.

impl<Interface, AddressType: Copy, Access> BufferOperation<'_, Interface, AddressType, Access>
where Interface: AsyncBufferInterface<AddressType = AddressType>, Access: ReadCapability,

pub async fn read_async( &mut self, buf: &mut [u8], ) -> Result<usize, Interface::Error>

Read some bytes from this source into the specified buffer, returning how many bytes were read.

Mirror function of embedded_io_async::Read::read.

pub async fn read_exact_async( &mut self, buf: &mut [u8], ) -> Result<(), ReadExactError<Interface::Error>>

Read the exact number of bytes required to fill buf.

This function calls read() in a loop until exactly buf.len() bytes have been read, waiting if needed.

Mirror function of embedded_io_async::Read::read_exact.