device_driver

Struct RegisterOperation

pub struct RegisterOperation<'i, Interface, AddressType: Copy, Register: FieldSet, Access> { /* private fields */ }

Object that performs actions on the device in the context of a register

Implementations

impl<Interface, AddressType: Copy, Register: FieldSet, Access> RegisterOperation<'_, Interface, AddressType, Register, Access>
where Interface: RegisterInterface<AddressType = AddressType>, Access: WriteCapability,

pub fn write<R>( &mut self, f: impl FnOnce(&mut Register) -> R, ) -> Result<R, Interface::Error>

Write to the register.

The closure is given the write object initialized to the reset value of the register. If no reset value is specified for this register, this function is the same as Self::write_with_zero.

pub fn write_with_zero<R>( &mut self, f: impl FnOnce(&mut Register) -> R, ) -> Result<R, Interface::Error>

Write to the register.

The closure is given the write object initialized to all zero.

impl<Interface, AddressType: Copy, Register: FieldSet, Access> RegisterOperation<'_, Interface, AddressType, Register, Access>
where Interface: RegisterInterface<AddressType = AddressType>, Access: ReadCapability,

pub fn read(&mut self) -> Result<Register, Interface::Error>

Read the register from the device

impl<Interface, AddressType: Copy, Register: FieldSet, Access> RegisterOperation<'_, Interface, AddressType, Register, Access>
where Interface: RegisterInterface<AddressType = AddressType>, Access: ReadCapability + WriteCapability,

pub fn modify<R>( &mut self, f: impl FnOnce(&mut Register) -> R, ) -> Result<R, Interface::Error>

Modify the existing register value.

The register is read, the value is then passed to the closure for making changes. The result is then written back to the device.

impl<Interface, AddressType: Copy, Register: FieldSet, Access> RegisterOperation<'_, Interface, AddressType, Register, Access>
where Interface: AsyncRegisterInterface<AddressType = AddressType>, Access: WriteCapability,

pub async fn write_async<R>( &mut self, f: impl FnOnce(&mut Register) -> R, ) -> Result<R, Interface::Error>

Write to the register.

The closure is given the write object initialized to the reset value of the register. If no reset value is specified for this register, this function is the same as Self::write_with_zero.

pub async fn write_with_zero_async<R>( &mut self, f: impl FnOnce(&mut Register) -> R, ) -> Result<R, Interface::Error>

Write to the register.

The closure is given the write object initialized to all zero.

impl<Interface, AddressType: Copy, Register: FieldSet, Access> RegisterOperation<'_, Interface, AddressType, Register, Access>
where Interface: AsyncRegisterInterface<AddressType = AddressType>, Access: ReadCapability,

pub async fn read_async(&mut self) -> Result<Register, Interface::Error>

Read the register from the device

impl<Interface, AddressType: Copy, Register: FieldSet, Access> RegisterOperation<'_, Interface, AddressType, Register, Access>
where Interface: AsyncRegisterInterface<AddressType = AddressType>, Access: ReadCapability + WriteCapability,

pub async fn modify_async<R>( &mut self, f: impl FnOnce(&mut Register) -> R, ) -> Result<R, Interface::Error>

Modify the existing register value.

The register is read, the value is then passed to the closure for making changes. The result is then written back to the device.